Test did not compile on checkout as is. Simple fix to explicitly retrieve primitive double to allow workspace compilation.

+++ uk.ac.gda.core.test/src/gda/util/QuantityFactoryTest.java
@@ -504,7 +504,7 @@ public class QuantityFactoryTest {
 		value = "345";
 		Integer i = new Integer(value);
 		o = i;
-		q = Quantity.valueOf(Integer.valueOf(value), SI.MILLI(SI.METER));
+		q = Quantity.valueOf(Integer.valueOf(value).doubleValue(), SI.MILLI(SI.METER));
 		assertTrue("quantity factory length not at expected value " + value, q.equals(QuantityFactory.createFromObject(
 				o, SI.MILLI(SI.METER))));
